Refactor - move Product aggregate to separate folder. Add IAggregateRoot marker interface.
