+++ src/SampleProject.API/Orders/ChangeCustomerOrder/ChangeCustomerOrderCommandHandler.cs
@@ -3,6 +3,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using MediatR;
 using SampleProject.Domain.Customers.Orders;
+using SampleProject.Domain.Products;
 
 namespace SampleProject.API.Orders.ChangeCustomerOrder
 {
