+++ src/SampleProject.Domain/Products/Product.cs
@@ -1,11 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using SampleProject.Domain.Customers.Orders;
 using SampleProject.Domain.SeedWork;
 
-namespace SampleProject.Domain.Customers.Orders
+namespace SampleProject.Domain.Products
 {
-    public class Product : Entity
+    public class Product : Entity, IAggregateRoot
     {
         public Guid Id { get; private set; }
