+++ src/SampleProject.Domain/SeedWork/IAggregateRoot.cs
@@ -1,0 +1,7 @@
+﻿namespace SampleProject.Domain.SeedWork
+{
+    public interface IAggregateRoot
+    {
+        
+    }
+}
