+++ src/SampleProject.Infrastructure/Orders/ProductEntityTypeConfiguration.cs
@@ -2,6 +2,7 @@
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using SampleProject.Domain.Customers.Orders;
+using SampleProject.Domain.Products;
 
 namespace SampleProject.Infrastructure.Orders
 {
