+++ src/SampleProject.Infrastructure/Orders/ProductRepository.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using SampleProject.Domain.Customers.Orders;
+using SampleProject.Domain.Products;
 
 namespace SampleProject.Infrastructure.Orders
 {
